+++ opo/tophat/defect/ode_path.f90
@@ -2,6 +2,6 @@ module ode_path
   use nrtype
   integer :: nok,nbad,kount
   logical :: save_steps=.true.
-  real(8) :: dxsav_rk, dxsav_sp
+  real(dp) :: dxsav_rk, dxsav_sp
 end module ode_path
