+++ opo/tophat/defect/rk_adaptive.f90
@@ -73,7 +73,7 @@ contains
   subroutine derivs(x,y,dydx)
     implicit none
 
-    real(8), INTENT(IN) :: x
+    real(dp), INTENT(IN) :: x
     complex(8), DIMENSION(:,:,:), INTENT(IN) :: y
     complex(8), DIMENSION(:,:,:), INTENT(OUT) :: dydx
  
@@ -119,13 +119,13 @@ contains
     IMPLICIT NONE    
 
     complex(8), DIMENSION(:,:,:), INTENT(INOUT) :: ystart    
-    real(8), INTENT(IN) :: x1,x2,eps,h1,hmin    
+    real(dp), INTENT(IN) :: x1,x2,eps,h1,hmin    
   
-    real(8), PARAMETER :: TINY=1.0e-30    
+    real(dp), PARAMETER :: TINY=1.0e-30    
     integer, PARAMETER :: MAXSTP=1000000000
 
     integer :: nstp    
-    real(8) :: h,hdid,hnext,x,xsav    
+    real(dp) :: h,hdid,hnext,x,xsav    
     complex(8), DIMENSION(size(ystart,1),size(ystart,2),size(ystart,3)) :: dydx,y,yscal    
 
     x=x1    
@@ -168,7 +168,7 @@ contains
     write(*,*) "too many steps in odeint"    
   CONTAINS    
     SUBROUTINE save_a_step_rk    
-      real(8) :: sx,sy    
+      real(dp) :: sx,sy    
       integer :: ix, iy    
       write(label,FMT="(i3)") kount    
       open(unit=22, file="phcplx-opo_spc"//trim(adjustl(label))//".dat", status='replace')    
@@ -197,9 +197,9 @@ contains
   SUBROUTINE odeint_sp(ystart,x1,x2,eps,h1,hmin)
     IMPLICIT NONE    
     complex(8), DIMENSION(:,:,:), INTENT(INOUT) :: ystart    
-    real(8), INTENT(IN) :: x1,x2,eps,h1,hmin    
+    real(dp), INTENT(IN) :: x1,x2,eps,h1,hmin    
   
-    real(8), PARAMETER :: TINY=1.0e-30    
+    real(dp), PARAMETER :: TINY=1.0e-30    
     integer, PARAMETER :: MAXSTP=1000000000
     !Runge-Kutta driver with adaptive step size control.Integrate the array
     !of starting values ystart from x1 to x2 with accuracy eps storing
@@ -211,7 +211,7 @@ contains
     !he stepper routine to be used.
     
     integer :: nstp    
-    real(8) :: h,hdid,hnext,x,xsav    
+    real(dp) :: h,hdid,hnext,x,xsav    
     complex(8), DIMENSION(size(ystart,1),size(ystart,2),size(ystart,3)) :: dydx,y,yscal    
     x=x1    
     h=sign(h1,x2-x1)    
@@ -271,9 +271,9 @@ contains
     IMPLICIT NONE    
     complex(8), DIMENSION(:,:,:), INTENT(INOUT) :: y    
     complex(8), DIMENSION(:,:,:), INTENT(IN) :: dydx,yscal    
-    real(8), INTENT(INOUT) :: x    
-    real(8), INTENT(IN) :: htry,eps    
-    real(8), INTENT(OUT) :: hdid,hnext    
+    real(dp), INTENT(INOUT) :: x    
+    real(dp), INTENT(IN) :: htry,eps    
+    real(dp), INTENT(OUT) :: hdid,hnext    
     INTEGER, EXTERNAL :: assert_eq    
     !Fifth order Runge-Kutta step with monitoring of local truncation error    
     !to ensure accuracy and adjust stepsize. Input are the dependent variable    
@@ -285,9 +285,9 @@ contains
     !accomplished, and hnext is the estimated next stepsize. derivs is the    
     !user-supplied subroutine that computes the right-hand-side derivatives.    
     integer :: ndum    
-    real(8) :: errmax,h,htemp,xnew    
+    real(dp) :: errmax,h,htemp,xnew    
     complex(8), DIMENSION(size(y,1),size(y,2),size(y,3)) :: yerr,ytemp    
-    real(8), PARAMETER :: SAFETY=0.9,PGROW=-0.2,PSHRNK=-0.25,&    
+    real(dp), PARAMETER :: SAFETY=0.9,PGROW=-0.2,PSHRNK=-0.25,&    
     ERRCON=1.89e-4    
     !The value ERRCON equals (5/SAFETY)**(1/PGROW),see use below.    
     ndum=assert_eq(size(y),size(dydx),size(yscal),'rkqs')     
@@ -323,7 +323,7 @@ contains
   SUBROUTINE rkck(y,dydx,x,h,yout,yerr)
     IMPLICIT NONE
     complex(8), DIMENSION(:,:,:), INTENT(IN) :: y,dydx
-    real(8), INTENT(IN) :: x,h
+    real(dp), INTENT(IN) :: x,h
     complex(8), DIMENSION(:,:,:), INTENT(OUT) :: yout,yerr
     INTEGER, EXTERNAL :: assert_eq
     !Given values for N variables y and their derivatives dydx known at x
@@ -335,7 +335,7 @@ contains
 
     integer :: ndum
     complex(8), DIMENSION(size(y,1),size(y,2),size(y,3)) :: ak2,ak3,ak4,ak5,ak6,ytemp
-    real(8), PARAMETER :: A2=0.2,A3=0.3,A4=0.6,A5=1.0,&
+    real(dp), PARAMETER :: A2=0.2,A3=0.3,A4=0.6,A5=1.0,&
              A6=0.875,B21=0.2,B31=3.0/40.0,B32=9.0/40.0,&
              B41=0.3,B42=-0.9,B43=1.2,B51=-11.0/54.0,&
              B52=2.5,B53=-70.0/27.0,B54=35.0/27.0,&
